fix: double pass to support syscalls on unfinished forks
Actual use case:

```
[pid 338627]      0.003957 vfork( <unfinished ...>
[pid 338628]      0.000068 execve("/usr/share/Modules/bin/cc", ...
[pid 338628]      0.000089 execve("/usr/local/bin/cc", ...
[pid 338628]      0.000022 execve("/usr/bin/cc", ...
[pid 338627]      0.000071 <... vfork resumed>) = 338628
```

Here the vfork ends and returns a new pid but execves are dones on this
new process in between its start and end, making line by line generation
impossible without some kind of kind of backlog.
When the process used on the left was not yet assigned to a fork, this
script was assuming it was the root pid, which was messing replacing the
root node incorrectly.